structure js for a bit of sanity
