+++ web/static/js/Home.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class Home extends Component {
+  render() {
+    return (
+      <div>
+        <p>Hello World!</p>
+      </div>
+    );
+  }
+}
+
+export default Home;
