+++ web/static/js/reducers/index.js
@@ -1,0 +1,5 @@
+const reducer = function(initialState = {}, state) {
+  return state;
+}
+
+export default reducer;
